move python to its own folder
