+++ Python/requirements.txt
@@ -1,0 +1,5 @@
+PyYAML
+pandas
+numpy scipy
+statsmodels
+datetime
