+++ Python/src/dataset.py
@@ -1,7 +1,7 @@
 import yaml
 import pandas as pd
 import numpy as np
-from Python.util import poly, dt_to_dec
+from src.util import poly, dt_to_dec
 from scipy.stats import gamma as gamma_scipy
 from numpy.random import gamma as gamma_np
 from statsmodels.distributions.empirical_distribution import ECDF
@@ -37,11 +37,11 @@ class HierarchicalDataset:
 
     def __init__(
         self,
-        config_dir="data/catalog.yml",
-        cases_dir="data/COVID-19-up-to-date.csv",
-        ifr_dir="data/weighted_fatality.csv",
-        serial_interval_dir="data/serial_interval.csv",
-        interventions_dir="data/interventions.csv",
+        config_dir="../data/catalog.yml",
+        cases_dir="../data/COVID-19-up-to-date.csv",
+        ifr_dir="../data/weighted_fatality.csv",
+        serial_interval_dir="../data/serial_interval.csv",
+        interventions_dir="../data/interventions.csv",
         num_countries=11,
         num_covariates=6,
         N2=75,
